requirements.txt
azure-identity==1.10.0
requests==2.31.0
pandas==2.0.3
